Refactor keymap_derive
